session: separate authprovider to package
