+++ client/solve.go
@@ -12,7 +12,6 @@ import (
 	"github.com/moby/buildkit/client/llb"
 	"github.com/moby/buildkit/identity"
 	"github.com/moby/buildkit/session"
-	"github.com/moby/buildkit/session/auth"
 	"github.com/moby/buildkit/session/filesync"
 	"github.com/moby/buildkit/session/grpchijack"
 	"github.com/moby/buildkit/solver/pb"
@@ -33,7 +32,7 @@ type SolveOpt struct {
 	FrontendAttrs     map[string]string
 	ExportCache       string
 	ImportCache       string
-	// Session string
+	Session           []session.Attachable
 }
 
 // Solve calls Solve on the controller.
@@ -76,7 +75,9 @@ func (c *Client) Solve(ctx context.Context, def *llb.Definition, opt SolveOpt, s
 		s.Allow(filesync.NewFSSyncProvider(syncedDirs))
 	}
 
-	s.Allow(auth.NewDockerAuthProvider())
+	for _, a := range opt.Session {
+		s.Allow(a)
+	}
 
 	switch opt.Exporter {
 	case ExporterLocal:
