+++ session/auth/auth.go
@@ -2,46 +2,11 @@ package auth
 
 import (
 	"context"
-	"io/ioutil"
 
-	"github.com/docker/cli/cli/config"
-	"github.com/docker/cli/cli/config/configfile"
 	"github.com/moby/buildkit/session"
-	netcontext "golang.org/x/net/context"
-	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
-
-func NewDockerAuthProvider() session.Attachable {
-	return &authProvider{
-		config: config.LoadDefaultConfigFile(ioutil.Discard),
-	}
-}
-
-type authProvider struct {
-	config *configfile.ConfigFile
-}
-
-func (ap *authProvider) Register(server *grpc.Server) {
-	RegisterAuthServer(server, ap)
-}
-
-func (ap *authProvider) Credentials(ctx netcontext.Context, req *CredentialsRequest) (*CredentialsResponse, error) {
-	if req.Host == "registry-1.docker.io" {
-		req.Host = "https://index.docker.io/v1/"
-	}
-	ac, err := ap.config.GetAuthConfig(req.Host)
-	if err != nil {
-		return nil, err
-	}
-	res := &CredentialsResponse{}
-	if ac.IdentityToken != "" {
-		res.Secret = ac.IdentityToken
-	} else {
-		res.Username = ac.Username
-		res.Secret = ac.Password
-	}
-	return res, nil
-}
 
 func CredentialsFunc(ctx context.Context, c session.Caller) func(string) (string, string, error) {
 	return func(host string) (string, string, error) {
@@ -51,6 +16,9 @@ func CredentialsFunc(ctx context.Context, c session.Caller) func(string) (string
 			Host: host,
 		})
 		if err != nil {
+			if st, ok := status.FromError(err); ok && st.Code() == codes.Unimplemented {
+				return "", "", nil
+			}
 			return "", "", err
 		}
 		return resp.Username, resp.Secret, nil
